Refactor the subcomponents under their own folder
