+++ client/src/pages/sniffers/SniffersPage.tsx
@@ -1,90 +1,12 @@
-import { useCallback, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import { SniffersSideBar } from "./SniffersSideBar";
 import { Sniffer, useSniffersStore } from "../../stores/sniffersStores";
 import { useAuthStore } from "../../stores/authStore";
 import { useSnackbar } from "../../hooks/useSnackbar";
-import { EndpointType, InvocationType } from "./types";
-import { EndpointSideBar } from "./EndpointSideBar";
-import { InvocationsBottomBar } from "./InvocationsBottomBar";
 import { LoadingIcon } from "./LoadingIcon";
-import { getEnpoints, getInvocations, getLiveInvocations } from "../../api/api";
-import { InvocationUpperBar } from "./InvocationUpperBar";
 import { useNavigate, useParams } from "react-router-dom";
-
-export const LivePage = () => {
-  const [invocations, setInvocations] = useState<InvocationType[]>([]);
-  const [loadingRequests, setLoadingRequests] = useState(false);
-  const { invocationId } = useParams();
-  const navigator = useNavigate();
-
-  const invocation = invocations.find((i) => i.id === invocationId);
-
-  const loadInvocations = async () => {
-    setLoadingRequests(true);
-    return getLiveInvocations()
-      .then((res) => {
-        const invocations = res.data;
-        setInvocations(invocations);
-        if (invocations.length > 0 && !invocationId) {
-          navigator(`/invocations/${invocations[0].id}`, {
-            replace: true,
-          });
-        }
-        return invocations;
-      })
-      .finally(() => {
-        setLoadingRequests(false);
-      });
-  };
-
-  useEffect(() => {
-    loadInvocations();
-    const int = setInterval(() => {
-      loadInvocations();
-    }, 60000);
-
-    return () => {
-      clearInterval(int);
-    };
-  }, []);
-
-  const onInvocationClick = useCallback(
-    (id: string) => {
-      navigator(`/invocations/${id}`);
-    },
-    [invocationId],
-  );
-
-  return (
-    <>
-      <div className={`flex flex-col w-full`}>
-        <div className="flex flex-col p-4 px-4 border-b border-border-color h-2/3 max-h-[calc(67vh-56px)] overflow-y-auto">
-          <InvocationUpperBar
-            activeInvocation={invocation}
-            activeSniffer={undefined}
-            onExecuteRequest={() => loadInvocations()}
-          />
-        </div>
-        <div className="flex flex-col p-2 px-4 h-1/3 max-h-[calc(33vh-16px)] overflow-y-auto overflow-x-auto">
-          {invocations &&
-            (loadingRequests ? (
-              <div className="flex flex-1 justify-center items-center">
-                <LoadingIcon />
-              </div>
-            ) : (
-              <InvocationsBottomBar
-                title={"Live Invocations"}
-                invocations={invocations}
-                activeInvocation={invocation}
-                setActiveInvocation={onInvocationClick}
-                refresh={() => loadInvocations()}
-              />
-            ))}
-        </div>
-      </div>
-    </>
-  );
-};
+import { SnifferData } from "./SniffersPage/SnifferData";
+import { LivePage } from "./SniffersPage/LivePage";
 
 const SniffersPage = () => {
   const { show: showSnackbar, component: snackBar } = useSnackbar();
@@ -136,133 +58,6 @@ const SniffersPage = () => {
         {sniffer ? <SnifferData sniffer={sniffer} /> : <LivePage />}
       </div>
     </div>
-  );
-};
-
-interface SnifferDataProps {
-  sniffer: Sniffer;
-}
-const SnifferData: React.FC<SnifferDataProps> = (props) => {
-  const navigator = useNavigate();
-  const { show: showSnackbar } = useSnackbar();
-  const [invocations, setInvocations] = useState<InvocationType[]>([]);
-  const [loadingRequests, setLoadingRequests] = useState(false);
-  const [endpoints, setEndpoints] = useState<EndpointType[]>([]);
-  const { snifferId, endpointId, invocationId } = useParams();
-  const [loadingEndpoints, setLoadingEndpoints] = useState(false);
-  const sniffer = props.sniffer;
-  const endpoint = endpoints.find((e) => e.id === endpointId);
-  const invocation = invocations.find((i) => i.id === invocationId);
-
-  // Invocations source function
-  const refreshInvocations = (endpointId: string) => {
-    setLoadingRequests(true);
-    getInvocations(endpointId)
-      .then((res) => {
-        setInvocations(res.data);
-        return res.data;
-      })
-      .then((invocations) => {
-        if (invocations.length > 0) {
-          navigator(
-            `/sniffers/${snifferId}/endpoints/${endpointId}/invocations/${invocations[0].id}`,
-            { replace: true },
-          );
-        }
-      })
-      .catch(() => {
-        setInvocations([]);
-        showSnackbar("Failed to get invocations", "error");
-      })
-      .finally(() => {
-        setLoadingRequests(false);
-      });
-  };
-
-  // populate the invocations
-  useEffect(() => {
-    if (!endpointId) {
-      setInvocations([]);
-    } else {
-      refreshInvocations(endpointId);
-    }
-  }, [endpointId]);
-
-  // Populate the endpoints of the screen
-  useEffect(() => {
-    if (!snifferId) return;
-    setLoadingEndpoints(true);
-    getEnpoints(snifferId)
-      .then((res) => {
-        setEndpoints(res.data);
-        if (res.data.length > 0) {
-          navigator(`/sniffers/${snifferId}/endpoints/${res.data[0].id}`, {
-            replace: true,
-          });
-        }
-      })
-      .catch(() => {
-        showSnackbar("Failed to get endpoints", "error");
-      })
-      .finally(() => {
-        setLoadingEndpoints(false);
-      });
-  }, [snifferId]);
-
-  const onEndpointClick = (endpointId: string) => {
-    setInvocations([]);
-    navigator(`/sniffers/${snifferId}/endpoints/${endpointId}`);
-  };
-
-  const onInvocationClick = (invocationId: string) => {
-    navigator(
-      `/sniffers/${snifferId}/endpoints/${endpointId}/invocations/${invocationId}`,
-    );
-  };
-
-  return (
-    <>
-      <div className={`flex flex-col w-[calc(100vw-296px-25%)]`}>
-        <div className="flex flex-col p-4 px-4 border-b border-border-color h-2/3 max-h-[calc(67vh-56px)] overflow-y-auto">
-          <InvocationUpperBar
-            activeInvocation={invocation}
-            activeSniffer={sniffer}
-            onExecuteRequest={() =>
-              endpointId && refreshInvocations(endpointId)
-            }
-          />
-        </div>
-        <div className="flex flex-col p-2 px-4 h-1/3 max-h-[calc(33vh-16px)] overflow-y-auto overflow-x-auto">
-          {invocations &&
-            (loadingRequests ? (
-              <div className="flex flex-1 justify-center items-center">
-                <LoadingIcon />
-              </div>
-            ) : (
-              <InvocationsBottomBar
-                title={"Invocations"}
-                invocations={invocations}
-                activeInvocation={invocation}
-                setActiveInvocation={onInvocationClick}
-                refresh={() => endpointId && refreshInvocations(endpointId)}
-              />
-            ))}
-        </div>
-      </div>
-      <div className="flex flex-col h-full max-h-[calc(100vh-96px)] min-w-[25%] p-4 border-l border-border-color overflow-y-auto">
-        {endpoints && loadingEndpoints ? (
-          <div className="flex flex-1 justify-center items-center">
-            <LoadingIcon />
-          </div>
-        ) : (
-          <EndpointSideBar
-            activeEndpoint={endpoint}
-            setActiveEndpoint={onEndpointClick}
-            requests={endpoints}
-          />
-        )}
-      </div>
-    </>
   );
 };
